Add Article model with related views, tests and database
This commit introduces an Article model to the application.
It includes all necessary files and changes to implement this model.
Migration file to create 'articles' table is added to the database, and the database schema is also updated.
In addition, related views under `app/views/articles` are included,
and a new controller 'ArticlesController' handles the necessary actions.
Database credentials are stored in the environmental variables using 'dotenv' gem.
Test files for both the Article model and the ArticlesController are in place, but still need to be populated with actual tests.
Routes have been updated to include ones for articles resources and for articles index page to be the root page.
To keep sensitive data safe, 'dotenv-rails' is added to the Gemfile for managing environmental variables.