Refactor views to extract comments functionality
This commit moves the comments section from the 'show' view of articles into its own partial views to promote code reusability. The form used to create a new comment has been moved to a '_form' partial in the comments view folder and the display of each comment has been moved to a '_comment' partial.